+++ addons/main/config.cpp
@@ -10,7 +10,9 @@ class CfgPatches {
             "wb_pb_prep"
         };
         author = "johnb43";
-        authors[] = {"johnb43"};
+        authors[] = {
+            "johnb43"
+        };
         url = "https://github.com/johnb432/Weapons-Balance-PB";
         VERSION_CONFIG;
     };