+++ addons/prep/config.cpp
@@ -17,7 +17,9 @@ class CfgPatches {
             "tfar_backpacks"
         };
         author = "johnb43";
-        authors[] = {"johnb43"};
+        authors[] = {
+            "johnb43"
+        };
         url = "https://github.com/johnb432/Weapons-Balance-PB";
         VERSION_CONFIG;
     };