+++ addons/prep/script_component.hpp
@@ -1,11 +1,10 @@
-#include "..\script_mod.hpp"
-
 #define COMPONENT prep
-#define COMPONENT_NAME QUOTE(Weapons Balance - Project Blufor Prep)
+#define COMPONENT_BEAUTIFIED Project-BLUFOR
+#include "..\main\script_mod.hpp"
 
 // #define DEBUG_MODE_FULL
 // #define DISABLE_COMPILE_CACHE
-// #define CBA_DEBUG_SYNCHRONOUS
+// #define DEBUG_SYNCHRONOUS
 // #define ENABLE_PERFORMANCE_COUNTERS
 
 #ifdef DEBUG_ENABLED_WB_PB_PREP
@@ -16,4 +15,4 @@
     #define DEBUG_SETTINGS DEBUG_SETTINGS_WB_PB_PREP
 #endif
 
-#include "\x\cba\addons\main\script_macros_common.hpp"
+#include "..\main\script_macros.hpp"