+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# Changelog for Weapons balance 8.10.2022
+# Changelog for Weapons balance 20.10.2023
+
+1.2.1.0
+- Minor fixes.
+
+# Changelog for Weapons balance 8.10.2023
 
 1.2.0.0
 - Minor cleanup.